Vision changes
When you move the right joystick, it will take a snapshot and try
to process it. It will save the color image and the processed image
to c_img.jpg and t_img.jpg.

DO NOT call Start() on BackboardFinder. There is some issue where writing
to a file from anything other than the FRC User Task causes
the crio to reboot.